test(factory): #329 split circuit-breaker unit tests into their own file (pure move)
